\name{arlab4-package}
\alias{arlab4-package}
\alias{arlab4}
\docType{package}
\title{
\packageTitle{arlab4}
}
\description{
\packageDescription{arlab4}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{arlab4}
\packageIndices{arlab4}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{arlab4}

Maintainer: \packageMaintainer{arlab4}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
