% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A multiple regression model (RC)}
\arguments{
\item{formula}{Contains dependent and independent variables for linear regression}

\item{data}{A data.frame to conduct linear regression}
}
\description{
Class for the convenient multiple linear regression
}
\section{Methods}{

\describe{
\item{\code{print()}}{Print out the coefficients and coefficient names}
}}

